created swagger.json created pagination for the receives the lists

+++ src/main/kotlin/ru/cgstore/routes/feedback/Resources.kt
@@ -14,7 +14,7 @@ class FeedBack {
     data class All(
         val parent: FeedBack = FeedBack(),
         val modelID: String = "",
-        val page: Int = 0,
+        val page: Long = 0,
         val size: Int = 16,
     )
 }
