+++ src/main/kotlin/ru/cgstore/routes/profile/Resources.kt
@@ -5,7 +5,7 @@ import io.ktor.resources.*
 @Resource("profile")
 class Profile {
     @Resource("models")
-    data class Models(val profile: Profile = Profile(), val page: Int = 1, val size: Int = 16)
+    data class Models(val profile: Profile = Profile(), val page: Long = 1, val size: Int = 16)
     @Resource("update")
     data class Update(val profile: Profile = Profile())
 }
